requirements.txt
atlassian-python-api==3.41.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
Deprecated==1.2.14
idna==3.4
msal==1.23.0
msal-extensions==1.0.0
mypy==1.5.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0

